requirements.txt
google-generativeai==0.3.2
pandas==2.2.1
scikit-learn==1.4.1.post1
gradio==4.24.0
openpyxl==3.1.2
python-dotenv==1.0.1